chore: move booleanSelectItems to where they're used
